Forgot to stage the previous commit. Nothing big although a small game over screen high score change

+++ catasstrophy/GameOverScreen.h
@@ -11,7 +11,7 @@
 
 @interface GameOverScreen : SKScene{
 @public
-    int highScore;
+    NSUInteger highScore;
     BOOL played;
 }
